Delete duplicated controllers, cleaning up
