+++ ItemManagment/Areas/Admin/Views/Shared/_Layout.cshtml
@@ -37,15 +37,14 @@
                 <a class="navbar-brand" href="/admin/home/index"> Admin Home </a>
 
                 <ul class="nav navbar-nav">
-                    <li> <a href="/admin/item/create"> Create Item </a> </li>
-                    <li> <a href="/admin/item/index"> Browse Items </a> </li>
-                    <li> <a href="/admin/itemCollection/create"> Create Collection </a> </li>
-                    <li> <a href="/admin/itemCollection/index"> Browse Collections </a> </li>
+                    <li> <a href="/item/create"> Create Item </a> </li>
+                    <li> <a href="/item/index"> Browse Items </a> </li>
+                    <li> <a href="/itemCollection/create"> Create Collection </a> </li>
+                    <li> <a href="/itemCollection/index"> Browse Collections </a> </li>
                     <li> <a href="/admin/applicationUser/index"> Manage Users </a> </li>
                     <li><a href="/account/myprofile"> My Profile </a></li>
                     <li><a href="/account/logout"> Log out </a></li>
                 </ul>
-
             </div>
         </div>
     </div>
