+++ ItemManagment/Controllers/ItemController.cs
@@ -1,10 +1,12 @@
-﻿using System;
+﻿using ItemManagement.DomainModels;
+using ItemManagement.DataLayer;
+using ItemManagement.ServiceLayer;
+using ItemManagement.Filters;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
-using ItemManagement.DomainModels;
-using ItemManagement.ServiceLayer;
 
 namespace ItemManagement.Controllers
 {
@@ -15,16 +17,55 @@ namespace ItemManagement.Controllers
         {
             _itemService = new ItemService();
         }
-        // GET: Item
         public ActionResult Index()
-        {
-            var items = _itemService.GetItems();
-            return View(items);
+        { 
+            List<Item> Items = _itemService.GetItems();
+
+            return View(Items);
         }
-        public ActionResult Details(long id)
+
+        [AuthenticationFilter]
+        public ActionResult Create()
         {
-            var item = _itemService.GetItemById(id);
+            return View();
+        }
+
+        [HttpPost]
+        public ActionResult Create(Item item)
+        {
+            if (item == null)
+            {
+                return View();
+            }
+            _itemService.CreateItem(item);
+
+            return RedirectToAction("Index");
+        }
+
+        [HttpGet]
+        public ActionResult Details(long Id)
+        {
+            var item = _itemService.GetItemById(Id);
             return View(item);
+        }
+        public ActionResult Edit(long id)
+        {
+            Item item = _itemService.GetItemById(id);
+            return View(item);
+        }
+
+        [HttpPost]
+        public ActionResult Edit(Item item)
+        {
+            _itemService.UpdateItem(item);
+            return RedirectToAction("Index");
+        }
+
+        
+        public ActionResult Delete(long id)
+        {
+            _itemService.DeleteItem(id);
+            return RedirectToAction("Index");
         }
     }
 }
